+++ packages/@ourworldindata/core-table/src/CoreTableUtils.ts
@@ -366,28 +366,6 @@ export const makeKeyFn =
             .map((slug) => toString(columnStore[slug][rowIndex]))
             .join(" ")
 
-// Memoization for immutable getters. Run the function once for this instance and cache the result.
-export const imemo = <Type>(
-    target: unknown,
-    propertyName: string,
-    descriptor: TypedPropertyDescriptor<Type>
-): void => {
-    const originalFn = descriptor.get!
-    descriptor.get = function (this: Record<string, Type>): Type {
-        const propName = `${propertyName}_memoized`
-        if (this[propName] === undefined) {
-            // Define the prop the long way so we don't enumerate over it
-            Object.defineProperty(this, propName, {
-                configurable: false,
-                enumerable: false,
-                writable: false,
-                value: originalFn.apply(this),
-            })
-        }
-        return this[propName]
-    }
-}
-
 export const appendRowsToColumnStore = (
     columnStore: CoreColumnStore,
     rows: CoreRow[]
