+++ packages/@ourworldindata/utils/src/index.ts
@@ -38,13 +38,14 @@ export {
     EnrichedBlockHtml,
     EnrichedBlockImage,
     EnrichedBlockList,
+    EnrichedBlockMissingData,
+    EnrichedBlockNumberedList,
     EnrichedBlockProminentLink,
     EnrichedBlockPullQuote,
     EnrichedBlockRecirc,
     EnrichedBlockScroller,
     EnrichedBlockSDGGrid,
     EnrichedBlockSDGToc,
-    EnrichedBlockMissingData,
     EnrichedBlockSideBySideContainer,
     EnrichedBlockSimpleText,
     EnrichedBlockStickyLeftContainer,
@@ -54,7 +55,6 @@ export {
     EnrichedRecircItem,
     EnrichedScrollerItem,
     EnrichedSDGGridItem,
-    EnrichedBlockNumberedList,
     EntryMeta,
     EntryNode,
     EPOCH_DATE,
@@ -101,24 +101,24 @@ export {
     RawBlockHtml,
     RawBlockImage,
     RawBlockList,
+    RawBlockMissingData,
+    RawBlockNumberedList,
     RawBlockPosition,
     RawBlockProminentLink,
     RawBlockPullQuote,
     RawBlockRecirc,
     RawBlockRecircValue,
     RawBlockScroller,
-    RawRecircItem,
     RawBlockSDGGrid,
+    RawBlockSDGToc,
     RawBlockSideBySideContainer,
     RawBlockStickyLeftContainer,
     RawBlockStickyRightContainer,
     RawBlockText,
     RawBlockUrl,
-    RawBlockNumberedList,
     RawChartStoryValue,
+    RawRecircItem,
     RawSDGGridItem,
-    RawBlockSDGToc,
-    RawBlockMissingData,
     BasicChartInformation,
     RelatedChart,
     ScaleType,
@@ -132,9 +132,9 @@ export {
     SpanFallback,
     SpanItalic,
     SpanLink,
-    SpanRef,
     SpanNewline,
     SpanQuote,
+    SpanRef,
     SpanSimpleText,
     SpanSubscript,
     SpanSuperscript,
@@ -143,6 +143,9 @@ export {
     SuggestedChartRevisionStatus,
     Tag,
     Time,
+    TimeBound,
+    TimeBounds,
+    TimeBoundValue,
     TimeRange,
     TocHeading,
     TocHeadingWithTitleSupertitle,
@@ -256,6 +259,9 @@ export {
     getArticleFromJSON,
     formatDate,
     canWriteToClipboard,
+    isNegativeInfinity,
+    isPositiveInfinity,
+    imemo,
 } from "./Util.js"
 
 export {
@@ -343,12 +349,7 @@ export {
 } from "./formatValue.js"
 
 export {
-    TimeBound,
-    TimeBounds,
-    TimeBoundValue,
     timeFromTimebounds,
-    isNegativeInfinity,
-    isPositiveInfinity,
     minTimeBoundFromJSONOrNegativeInfinity,
     maxTimeBoundFromJSONOrPositiveInfinity,
     minTimeToJSON,
